requirements.txt
boto3==1.18.0
psycopg2-binary==2.9.1

